fix: address React Doctor review comments
- CmsToggle: add aria-label for accessibility
- EditableField: remove derived useState, move focus logic from useEffect to event handler, add role=button + key handler for a11y
- EditableImage: use useRef instead of useState for loaded flag, add a11y attributes
- EditableSelect: add role=button + key handler, remove autoFocus
- QuestionCard: replace array index key with compound key, wrap checkbox in label
- cms.tsx: wrap context value in useMemo to avoid unstable references
- SubjectHome: extract ExamCardEditor component to reduce giant component size
