+++ src/components/CmsEditor/EditableSelect.tsx
@@ -17,6 +17,13 @@ export default function EditableSelect({
 
   const selectedOption = options.find((o) => o.value === value);
 
+  const handleDivKeyDown = (e: React.KeyboardEvent) => {
+    if (e.key === "Enter" || e.key === " ") {
+      e.preventDefault();
+      setEditing(true);
+    }
+  };
+
   return (
     <div>
       {label && (
@@ -28,7 +35,10 @@ export default function EditableSelect({
         <div
           className="group cursor-pointer rounded-md border border-dashed border-transparent hover:border-amber-400/60 hover:bg-amber-50/30 px-1 py-0.5 -mx-1 -my-0.5 transition-all text-sm"
           onClick={() => setEditing(true)}
-          title="Click to change"
+          onKeyDown={handleDivKeyDown}
+          role="button"
+          tabIndex={0}
+          aria-label={`Edit ${label ?? "selection"}`}
         >
           {selectedOption ? (
             <span>{selectedOption.label}</span>
@@ -45,7 +55,7 @@ export default function EditableSelect({
           }}
           onBlur={() => setEditing(false)}
           className="w-full p-2 text-sm rounded-md border-2 border-amber-400 bg-surface focus:border-amber-500 focus:ring-2 focus:ring-amber-500/30 focus:outline-none"
-          autoFocus
+          aria-label={label}
         >
           {options.map((o) => (
             <option key={o.value} value={o.value}>
